// src/gen-ui/lib/core/base-types.ts
export interface ImageStreamPayload<T = Uint8Array> {
  value: T;
  total: number;
  error: any;
  progress: number;
  done: boolean;
}
export type ImageLoadPayload = {
  size: Size;
  image: any;
};

export interface ImageProviderLoadConfiguration {
  url: string;
  headers?: Record<string, string>;
  lifecycle?: Partial<ImageProviderLifecycle>;
}

export interface ImageProviderLifecycle {
  onLoadStart: () => void;
  onProgress: (progress: number) => void;
  onLoadEnd: (imagePayLoad: ImageLoadPayload) => void;
  onError: (error: any) => void;
}